requirements.txt
Django==3.2.10
pillow
djangorestframework==3.13.1
